% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregacja_adm3.R
\name{agreguj_aneks_1rokpo_adm_2}
\alias{agreguj_aneks_1rokpo_adm_2}
\title{Obliczanie wskaznikow na poziomie zagregowanym dla 3. edycji
monitoringu na danych administracyjnych}
\usage{
agreguj_aneks_1rokpo_adm_2(wsk3, wsk4, podzial_grupy, rok_abso)
}
\arguments{
\item{wsk3}{ramka danych z tabeli pośredniej nr 3 (P3) z wynikami z 2.
rundy monitoringu na danych administracyjnych}

\item{wsk4}{ramka danych z tabeli pośredniej nr 4 (P4) z wynikami z 2.
rundy monitoringu na danych administracyjnych}

\item{podzial_grupy}{ramka danych zawierająca definicje podziałów na grupy -
np. zwrócona przez funkcję \code{\link{utworz_grupowanie_ze_zmiennej}}}

\item{rok_abso}{rok, w którym grupa absolwentów uzyskała status absolwenta.
Może to być więcej niż 1 wartość.}
}
\value{
data frame
}
\description{
Funkcja obliczająca wskaźniki na poziomie zagregowanym na
potrzeby Aneksu w raportach "1rokpo" - liczy mniej wskaźników niż
\code{\link{agreguj_1rokpo_adm_2}}
}
\seealso{
\code{\link{agreguj_wskazniki}} oraz przekazywane do niej funkcje
używane do obliczania konkretnych wskaźników zagregowanych:
\itemize{
\item{\code{\link{l_abs}},}
\item{\code{\link{l_kobiet}},}
\item{\code{\link{S3_mies}},}
\item{\code{\link{zawody_S3}},}
\item{\code{\link{E2_nauka_kontyn}},}
\item{\code{\link{Z4_ods_prac_mies}},}
\item{\code{\link{Z8_formy_prac_mies}},}
\item{\code{\link{B2_ods_bezrob}},}
\item{\code{\link{licz_zawody}}}
}
}
