% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_adm3.R
\name{B2_ods_bezrob}
\alias{B2_ods_bezrob}
\title{Obliczanie wskaźników dla 3. edycji monitoringu - dane administracyjne}
\usage{
B2_ods_bezrob(x, rok, od = 9, do = 12)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{rok}{rok lub zakres lat osiągnięcia statusu absolwenta}

\item{od}{początek okresu wyliczania wskaźnika wyrażony miesiącem}

\item{do}{koniec okresu wyliczania wskaźnika wyrażony miesiącem}
}
\value{
lista
}
\description{
Funkcja licząca na potrzeby szablonu raportu rozkład liczby
miesięcy bezrobocia rejestrowanego wśród absolwentów w danym okresie.
Domyślnie jest to okres między wrześniem a grudniem roku przekazanego w
argumentach funkcji, ale te wartości można zmieniać.
}
