% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_adm3.R
\name{dane_szkoly}
\alias{dane_szkoly}
\title{Obliczanie wskaźników dla 3. edycji monitoringu - dane administracyjne}
\usage{
dane_szkoly(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
lista
}
\description{
Funkcja przechowująca nazwę i adres szkoły, której dotyczy
raport.
}
