% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_adm3.R
\name{l_abs}
\alias{l_abs}
\title{Obliczanie wskaźników dla 3. edycji monitoringu - dane administracyjne}
\usage{
l_abs(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
liczba
}
\description{
Funkcja oblicza wskaźnik opisujący liczbę absolwentów na danym
poziomie agregacji.
}
