% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_adm3.R
\name{l_abs_zrodla}
\alias{l_abs_zrodla}
\title{Obliczanie wskaźników dla 3. edycji monitoringu - dane administracyjne}
\usage{
l_abs_zrodla(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
lista
}
\description{
Funkcja oblicza wskaźnik opisujący liczbę absolwentów na danym
poziomie agregacji, o których pozyskano informacje z poszczególnych
rejestrów.
}
