% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_adm3.R
\name{l_kobiet}
\alias{l_kobiet}
\title{Obliczanie wskaźników dla 3. edycji monitoringu - dane administracyjne}
\usage{
l_kobiet(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
lista
}
\description{
Funkcja oblicza wskaźnik opisujący liczbę kobiet wśród
absolwentów na danym poziomie agregacji, którzy zostali objęci monitoringiem.
Dodatkowo na podstawie liczby kobiet obliczana i przechowywana jest forma
odmiany słowa "kobieta" w zależności od ich liczby (na potrzeby raportów
automatycznych).
}
