% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_adm3.R
\name{liczebnosc_dyscypliny_plec}
\alias{liczebnosc_dyscypliny_plec}
\title{Obliczanie wskaźników dla 3. edycji monitoringu - dane administracyjne}
\usage{
liczebnosc_dyscypliny_plec(x, dyscyplina_kont_df, rok, mies = 12, plc)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{dyscyplina_kont_df}{ramka danych zawierająca informację o kontynuowaniu
kształcenia w danej dyscyplinie (tabela danych pośrednich P2 lub zawierająca
analogiczne informacje oraz te same nazwy kolumn co tabela P2)}

\item{rok}{rok lub zakres lat osiągnięcia statusu absolwenta}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik - domyślnie jest
to grudzień}

\item{plc}{płeć absolwenta przekazana jako wartość tekstowa ("K" lub "M")}
}
\value{
lista
}
\description{
Funkcja licząca rozkład liczebności absolwentów danej płci
kontynuujących naukę na studiach w podziale na dyscypliny. Wskaźnik liczony
jest tylko dla absolwentów techników i liceów ogólnokształcących.
}
