% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_adm3.R
\name{odmiany}
\alias{odmiany}
\title{Obliczanie wskaźników dla 3. edycji monitoringu - dane administracyjne}
\usage{
odmiany(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
lista
}
\description{
Funkcja oblicza i przechowuje odmiany wyrażeń tekstowych, które
zostaną użyte w raportach.
}
