% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_adm3.R
\name{S3_mies}
\alias{S3_mies}
\title{Obliczanie wskaźników dla 3. edycji monitoringu - dane administracyjne}
\usage{
S3_mies(x, rok_od, mies_od, rok_do, mies_do)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{rok_od}{rok początku okresu, dla którego ma być policzony wskaźnik}

\item{mies_od}{miesiąc początku okresu, dla którego ma być policzony wskaźnik}

\item{rok_do}{rok końca okresu, dla którego ma być policzony wskaźnik}

\item{mies_do}{miesiąc końća okresu, dla którego ma być policzony wskaźnik}
}
\value{
lista
}
\description{
Funkcja licząca odsetek absolwentów o danym statusie
edukacyjno-zawodowym (wskaźnik S3) w danym miesiącu \code{mies} dla raportu
dla danego roku \code{rok}. Wyróżniamy następujące statusy (bez KUZ i KKZ):
\itemize{
\item{ucz_prac - nauka i praca}
\item{tylko_ucz - tylko nauka}
\item{tylko_prac - tylko praca}
\item{neet - brak nauki i brak pracy lub brak informacji o tychże}
}
}
