% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_adm3.R
\name{W3_sr_doch_uop}
\alias{W3_sr_doch_uop}
\title{Obliczanie wskaźników dla 3. edycji monitoringu - dane administracyjne}
\usage{
W3_sr_doch_uop(x, rok, od = 9, do = 12, nauka)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{rok}{rok lub zakres lat osiągnięcia statusu absolwenta}

\item{od}{początek okresu wyliczania wskaźnika wyrażony miesiącem}

\item{do}{koniec okresu wyliczania wskaźnika wyrażony miesiącem}

\item{nauka}{wartość TRUE/FALSE określająca czy status ma być liczony dla
absolwentów uczących się czy nie uczących się}
}
\value{
lista
}
\description{
Funkcja licząca średni względny miesięczny przychód z pracy
\emph{etatowej} w danym okresie w odniesieniu do zarobków w zamieszkiwanym
powiecie. Domyślnie jest to okres między wrześniem a grudniem roku
przekazanego w argumentach funkcji, ale te wartości można zmieniać. Średnia
liczona jest oddzielnie dla uczących i nie uczących się absolwentów.
}
