% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_adm3.R
\name{zawody_S3}
\alias{zawody_S3}
\title{Obliczanie wskaźników dla 3. edycji monitoringu - dane administracyjne}
\usage{
zawody_S3(x, rok_od, mies_od, rok_do, mies_do)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{rok_od}{rok początku okresu, dla którego ma być policzony wskaźnik}

\item{mies_od}{miesiąc początku okresu, dla którego ma być policzony wskaźnik}

\item{rok_do}{rok końca okresu, dla którego ma być policzony wskaźnik}

\item{mies_do}{miesiąc końća okresu, dla którego ma być policzony wskaźnik}
}
\value{
lista
}
\description{
Funkcja licząca odsetek absolwentów o danym statusie
edukacyjno-zawodowym wśród grup absolwentów wydzielonych ze względu na zawód,
w którym się kształcili w szkole. Zwracana lista służy jako wsad do tabeli w
raporcie automatycznym. Jeśli lista będąca wsadem tabeli generowałaby pustą
tabelę, to zwracana jest pusta lista.
}
\details{
Funkcja wykorzystuje inną funkcję z pakietu, czyli
\code{\link{S3_mies}}
}
